refactor: bump serde, part 1
