+++ rmp-serde/.cargo/config
@@ -1,2 +1,2 @@
 [build]
-target-dir = "../../target"
+target-dir = "../target"
