+++ rmp-serde/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rmp-serde"
-version = "0.10.0"
+version = "0.11.0"
 authors = ["Evgeny Safronov <division494@gmail.com>"]
 license = "MIT"
 description = "Serde bindings for RMP"
@@ -10,6 +10,6 @@ readme = "../README.md"
 keywords = ["msgpack", "MessagePack", "serder", "serialization"]
 
 [dependencies]
-rmp = { version = "^0.7", path = "../rmp" }
+rmp = { version = "^0.8", path = "../rmp" }
 serde = "^0.8"
 serde_derive = { version = "^0.8", optional = true }
